+++ libs/tokens/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [3.8.3](https://github.com/cowprotocol/cowswap/compare/tokens-v3.8.2...tokens-v3.8.3) (2026-07-08)
+
+### Patch Changes
+
+- fix: add reverve protocol token list (#7824)
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.8.0
+  - @cowprotocol/common-const@3.6.0
+  - @cowprotocol/core@3.6.0
+  - @cowprotocol/common-hooks@3.5.1
+  - @cowprotocol/ui@3.9.1
+
 ## [3.8.2](https://github.com/cowprotocol/cowswap/compare/tokens-v3.8.1...tokens-v3.8.2) (2026-06-25)
 
 ### Patch Changes